requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
pydantic[email]
openai
